+++ pdf/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 3.4.1
 
 - Fix Nunito font parsing
+- Allow reusing an ImageProvider and Font on multiple documents
 
 ## 3.4.0
 
@@ -19,7 +20,7 @@
 
 ## 3.3.0
 
-- Implement To be signed flieds
+- Implement To be signed fields
 - Improve Text rendering
 - Add individual cell decoration
 - Improve Bullet Widget